# Improve test utility header compilation speed