This PR reduces physical dependencies between headers for test utilities, resulting in 1-2s average compilation speedup on my machine.