@ChrisThrasher 2023-02-20
`master` (without running tests)
```
real    0m7.925s
user    0m56.337s
sys     0m6.599s
```

This PR (without running tests)
```
real    0m7.952s
user    0m56.119s
sys     0m6.551s
```

I'm not able to observe a build time reduction with this PR.