@vittorioromeo 2023-02-20
> Looks mostly good -- I wonder if we should still define the vector typedefs, as we're using them in a few places? Just for consistency with the rest of the codebase.

You mean in `SystemUtil`, right? I'd rather avoid doing so, it's a bit of extra repetition that we need to maintain between the actual vector header and the test header.