@vittorioromeo 2023-02-20
SFMLBuildMaster: Build this please.