@vittorioromeo 2023-02-20
> I'm failing to demonstrate any compile time improvements and even if I could show that it was faster for me, I'm still not keen on adding complexity for marginal compile time gains that aren't even relevant to the core library. I won't veto this PR if others like it but I'm still not excited by it.

Here are some measurements on my machine, using `hyperfine "ninja clean; ninja"`, for a clean rebuild including all test drivers.

Before this PR:

```
  Time (mean ± σ):     15.025 s ±  0.531 s    [User: 120.985 s, System: 17.608 s]
  Range (min … max):   14.178 s … 15.642 s    10 runs
```

After this PR:

```
  Time (mean ± σ):     13.684 s ±  0.111 s    [User: 111.092 s, System: 16.656 s]
  Range (min … max):   13.537 s … 13.891 s    10 runs
```