@ChrisThrasher 2023-02-20
I was unaware of hyperfine. Looks pretty useful for these things. I again commented out the command that runs the tests since I don't want those skewing the results. (I'm growing sour on SFML running tests automatically after the build and might prefer we switch to he more idiomatic approach where tests are only run on demand. If we care about incremental build time improvements that's an easy way to improve).

`master`:
```
Benchmark 1: ninja clean; ninja
  Time (mean ± σ):      9.657 s ±  0.209 s    [User: 62.755 s, System: 7.995 s]
  Range (min … max):    9.424 s …  9.974 s    10 runs
```

This PR:
```
Benchmark 1: ninja clean; ninja
  Time (mean ± σ):      9.423 s ±  0.115 s    [User: 62.471 s, System: 7.878 s]
  Range (min … max):    9.303 s …  9.722 s    10 runs
```

The difference is still imperceptible to me.