@ChrisThrasher 2023-02-20
> CI failures seem to be related to codecov -- how to proceed?

https://github.com/ChrisThrasher/SFML/commit/5f8b7bf4f0c40fe039d6bf08836f446b4815da3f

Unrelated to this PR but I propose we allow coverage jobs to fail. I'd rather have a rock-solid CI pipeline than perfect code coverage information.