@vittorioromeo 2023-04-04
Rebased on `master`. @ChrisThrasher, are you still against this PR given Bromeon's observation? We are very likely to not touch this code again, but it will continuously give us benefits in terms of compilation speed and CI usage over the course of the next years.