@vittorioromeo 2023-04-13
@eXpl0it3r @ChrisThrasher: how do we proceed?

BTW, here's the latest benchmark after rebasing:

```
without PR:
Benchmark 1: ninja clean && ninja
  Time (mean ± σ):     16.227 s ±  0.570 s    [User: 169.491 s, System: 22.362 s]
  Range (min … max):   15.517 s … 17.353 s    10 runs

with PR:
Benchmark 1: ninja clean && ninja
  Time (mean ± σ):     15.539 s ±  0.481 s    [User: 161.124 s, System: 21.582 s]
  Range (min … max):   14.774 s … 16.092 s    10 runs
```